+++ tests/testthat/helper-endpoint.R
@@ -58,8 +58,7 @@ validate_baseline_all_input <- function(pjnz, shape, population) {
   )
 }
 
-validate_programme_survey_input <- function(file_path, type, shape,
-                                            pjnz) {
+validate_programme_survey_input <- function(file_path, type, shape) {
   sprintf(
     '{"type": "%s",
       "file": {
@@ -69,9 +68,8 @@ validate_programme_survey_input <- function(file_path, type, shape,
         "fromADR": false,
         "resource_url": "https://adr.unaids.org/file/123.csv"
       },
-      "shape": "%s",
-      "pjnz": "%s"
-    }', type, file_path, shape, pjnz)
+      "shape": "%s"
+    }', type, file_path, shape)
 }
 
 input_time_series_request <- function(file_path, type, shape_path) {