+++ tests/testthat/test-endpoints-validate-survey-and-programme.R
@@ -4,8 +4,7 @@ test_that("endpoint_validate_survey_programme supports programme file", {
   input <- validate_programme_survey_input(
     file.path("testdata", "programme.csv"),
     "programme",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(response$filename, scalar("original"))
@@ -21,8 +20,7 @@ test_that("endpoint_validate_survey_programme returns error on invalid programme
   input <- validate_programme_survey_input(
     file.path("testdata", "malformed_programme.csv"),
     "programme",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   error <- expect_error(validate_survey_programme(input))
 
   expect_equal(error$data[[1]]$error, scalar("INVALID_FILE"))
@@ -36,8 +34,7 @@ test_that("endpoint_validate_survey_programme supports ANC file", {
   input <- validate_programme_survey_input(
     file.path("testdata", "anc.csv"),
     "anc",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(response$filename, scalar("original"))
@@ -54,8 +51,7 @@ test_that("endpoint_validate_survey_programme returns error on invalid ANC data"
   input <- validate_programme_survey_input(
     file.path("testdata", "malformed_anc.csv"),
     "anc",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   error <- expect_error(validate_survey_programme(input))
 
   expect_equal(error$data[[1]]$error, scalar("INVALID_FILE"))
@@ -71,8 +67,7 @@ test_that("endpoint_validate_survey_programme supports survey file", {
   input <- validate_programme_survey_input(
     file.path("testdata", "survey.csv"),
     "survey",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(response$filename, scalar("original"))
@@ -88,8 +83,7 @@ test_that("endpoint_validate_survey_programme returns error on invalid survey da
   input <- validate_programme_survey_input(
     file.path("testdata", "malformed_survey.csv"),
     "survey",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   error <- expect_error(validate_survey_programme(input))
 
   expect_equal(error$data[[1]]$error, scalar("INVALID_FILE"))
@@ -103,8 +97,7 @@ test_that("possible filters are returned for data", {
   input <- validate_programme_survey_input(
     file.path("testdata", "programme.csv"),
     "programme",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(names(response$filters), c("age", "calendar_quarter", "indicators"))
@@ -141,8 +134,7 @@ test_that("possible filters are returned for data", {
   input <- validate_programme_survey_input(
     file.path("testdata", "anc.csv"),
     "anc",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(names(response$filters), c("year", "indicators"))
@@ -159,8 +151,7 @@ test_that("possible filters are returned for data", {
   input <- validate_programme_survey_input(
     file.path("testdata", "survey.csv"),
     "survey",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(names(response$filters), c("age", "surveys", "indicators"))
@@ -202,8 +193,7 @@ test_that("filters not returned if indicator missing from input data", {
   input <- validate_programme_survey_input(
     file.path("testdata", "programme_no_vls.csv"),
     "programme",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   response <- validate_survey_programme(input)
 
   expect_equal(names(response$filters),
@@ -333,8 +323,7 @@ test_that("anc data can be validated can be run with relaxed validation", {
   input <- validate_programme_survey_input(
     t,
     "anc",
-    file.path("testdata", "malawi.geojson"),
-    file.path("testdata", "Malawi2019.PJNZ"))
+    file.path("testdata", "malawi.geojson"))
   queue <- test_queue(workers = 0)
   api <- api_build(queue)